refactor: move LoginAuthVoArgumentResolver, ErrorCode to other packages
